Vision Code that Should not be on This Branch
This commit is for vision code that shouldn't be on this branch but is
due to lack of ability to change branches. We have added multiple things
to do with vision. Vision is needed and working to some extent. Tracking
is not yet up, but vision processing is somewhat working.