One Controller
It is a bit Hackish, but M can fix later.

As we stand, you could run the robot for demo now.
There is outstanding to make the turret rotate.  But for now I am saying
we are good for SIT.  Just need a router and to get it working there.